controllers

+++ database/migrations/2025_05_14_191030_create_employers_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
 
 
             $table->id();
-            $table->Integer('role_id')->unique();
+            $table->unsignedBigInteger('role_id')->unique();
             $table->foreign('role_id')->references('id')->on('users')->onDelete('cascade');
             $table->string('company_name');
             $table->text('description')->nullable();
